% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HJSTATICO.R
\name{HJSTATICO}
\alias{HJSTATICO}
\title{Multivariate Analysis Of Multi-Way Data Matrices}
\usage{
HJSTATICO(x, y, condition, biplot, preprocessing, compx, compy)
}
\arguments{
\item{x}{set of matrices juxtaposed in vertical order containing the data of the P variables.}

\item{y}{set of matrices juxtaposed in vertical order containing the data of the Q variables.}

\item{condition}{A factor object representing the categorical variable that will define the third-path for the analysis.}

\item{biplot}{Logical value, TRUE by default, that will calculate and graphically represent the and intra-structure considering the HJ-Biplot perspective. HJ-Biplot perspective.}

\item{preprocessing}{Standardization method to be used. preprocessing="B.Total" performs total standardization, preprocessing="B.Partial" partial standardization, preprocessing="A.Norma" standardization by standard.}

\item{compx}{Indicates the component of x to be used for the representation of the compromise and intra-structure plots on the bi-factor plane.}

\item{compy}{Indicates the component of y to be used for the representation of the compromise and intra-structure plots on the bi-factor plane.}
}
\value{
A series of indices (contributions and representation qualities) and compromise and intra-structure plots
}
\description{
Multivariate Analysis Of Multi-Way Data Matrices
}
\examples{
result=HJSTATICO(x=env_data,y=spe_data,condition=etiq_data$space,biplot=TRUE,preprocessing = "A.Norma",compx=1,compy=2)
result$Compromise
result$HJB
result$HJBK
result$CrossTables;result$X;result$Y
}
