% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prep.Tables.R
\name{Prep.Tables}
\alias{Prep.Tables}
\title{Preprocessing of tables: centering and normalizing}
\usage{
Prep.Tables(x, y, condition, preprocessing)
}
\arguments{
\item{x}{set of matrices juxtaposed in vertical order containing the data of the P variables.}

\item{y}{set of matrices juxtaposed in vertical order containing the data of the Q variables.}

\item{condition}{A factor object representing the categorical variable that will define the third-path for the analysis.}

\item{preprocessing}{Standardization method to be used. preprocessing="B.Total" performs total standardization, preprocessing="B.Partial" partial standardization, preprocessing="A.Norma" standardization by standard.}
}
\value{
Processed table and the condition index
}
\description{
Preprocessing of tables: centering and normalizing
}
\examples{
Prep.Tables(x=env_data,y=spe_data,condition=etiq_data$space,preprocessing = "A.Norma")
}
